Add remarks about complexity and copyless prop
In main decision procedure section, mention that the reduction to PSSTs
only uses copyless PSST without $<- and $->. Then in the complexity
analysis, state that we have a lower-bound for the tower-hardness, but
when there is no $<- or $->, then it's copyless and the tower of
exponentials does not happen.

In the appendix, when doing the pre-image, it was remarked that the
decision procedure does not need the copyless property. Expand on this
to explain that when the PSST is copyless, the abstraction does not need
a powerset and is not exponential in the input A.