Tidy up modelling as PSST section and put details back in appendix
Fixed a few typos and phrasing in the discussion in the main body of
modelling replaceAll.

Reinstated in the appendix the encoding of extract and the formal
details of replaceAll. I felt that the intuitive explanations were too
sketchy for an appendix and since the formal details were typed up, we
should just use them.