Handle converting x \in e to x \in A
Make explicit in reduction that we can just build an FA because
greediness and capture groups do not change the language accepted.

Clarify the need for an effective representation of RegEx constraints
in calculus. I.e. that it means membership constraints, and with a
footnote reminding the reader that we can just use standard techniques
for them.